requirements.txt
# Python dependencies
flask>=3.0
python-docx>=1.1
lxml>=5.0

# System dependency (install separately):
#   apt-get install libreoffice-writer
